# shared/contracts/openapi.nexus-core.snapshot.yaml
openapi: 3.0.3
info:
  title: Nexus Core API
  version: "1.1.0"
  description: |
    Multi-tenant Agent Tool Gateway for controlled tool execution.

    Authentication supports API key (`X-NEXUS-GATEWAY-KEY`) and optional JWT bearer mode (`Authorization: Bearer`) when enabled by feature flags.

    Idempotency semantics for `POST /v1/run` on WRITE tools:
    - Same `Idempotency-Key` + same request fingerprint + `COMPLETED` => replay cached success.
    - Same `Idempotency-Key` + same fingerprint + `IN_PROGRESS` => `IDEMPOTENCY_IN_PROGRESS`.
    - Same `Idempotency-Key` + different fingerprint => `IDEMPOTENCY_KEY_CONFLICT`.
    - Same `Idempotency-Key` + same fingerprint + `FAILED` => terminal replay of cached error.
    - To retry after `FAILED`, clients MUST use a new `Idempotency-Key`.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/tools:
    get:
      tags: [Tools]
      operationId: listTools
      summary: List tenant tools
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XNexusActor"
        - $ref: "#/components/parameters/XNexusRole"
        - $ref: "#/components/parameters/XNexusScopes"
      responses:
        "200":
          description: Tools list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Tools]
      operationId: createTool
      summary: Create tool
      security:
        - NexusApiKey: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolCreateRequest"
      responses:
        "201":
          description: Tool created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "400":
          description: Validation or schema error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/run:
    post:
      tags: [Run]
      operationId: runTool
      summary: Run a tool through Nexus policy pipeline
      description: |
        Executes: tool resolution -> schema validation -> policy/limits -> DLP/egress -> HTTP executor -> audit append.

        Idempotency applies to WRITE tools (non-GET by fallback).
        `FAILED` idempotency records are terminal: repeated same key/fingerprint returns cached error.
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XNexusActor"
        - $ref: "#/components/parameters/XNexusRole"
        - $ref: "#/components/parameters/XNexusScopes"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/TimeoutMs"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: Allowed and executed successfully.
          headers:
            X-Idempotency-Outcome:
              schema:
                type: string
                enum: [NEW, REPLAY, SKIPPED_NOT_WRITE]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunSuccessResponse"
        "400":
          description: Validation/policy requirement errors.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/RunBlockedResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
              examples:
                idempotencyRequired:
                  value:
                    request_id: "b6a0fa23-58a4-41ea-a45e-1a45f9c3f0e1"
                    decision: deny
                    status: blocked
                    reason: idempotency key required by policy
                    error:
                      code: IDEMPOTENCY_REQUIRED
                      message: idempotency key required by policy
                    latency_ms: 4
        "403":
          description: Blocked by policy/egress/rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunBlockedResponse"
        "408":
          description: Timeout budget exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunErrorResponse"
        "409":
          description: Idempotency conflict or in-progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunBlockedResponse"
              examples:
                keyConflict:
                  value:
                    request_id: "e6372db4-0cf4-43e9-b31f-6ebd1f8a0463"
                    decision: deny
                    status: blocked
                    reason: idempotency key used with different payload
                    error:
                      code: IDEMPOTENCY_KEY_CONFLICT
                      message: idempotency key used with different payload
                    latency_ms: 1
                inProgress:
                  value:
                    request_id: "f6b2e850-ad19-4bd4-ab20-4f2d98db3b57"
                    decision: deny
                    status: blocked
                    reason: idempotency request in progress
                    error:
                      code: IDEMPOTENCY_IN_PROGRESS
                      message: idempotency request in progress
                    latency_ms: 1
        "502":
          description: Upstream/network/output schema execution errors.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunErrorResponse"
              examples:
                failedReplay:
                  value:
                    request_id: "6fa0e110-df8f-4277-a4e6-4d9f2a3200e1"
                    decision: allow
                    status: error
                    error:
                      code: UPSTREAM_5XX
                      message: upstream server returned 5xx
                    latency_ms: 2
                    idempotency:
                      present: true
                      outcome: REPLAY
  /v1/run/simulate:
    post:
      tags: [Run]
      operationId: simulateRun
      summary: Simulate run decision/explain without upstream execution
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XNexusActor"
        - $ref: "#/components/parameters/XNexusRole"
        - $ref: "#/components/parameters/XNexusScopes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateRequest"
      responses:
        "200":
          description: Simulation result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulateResponse"
        "403":
          description: Denied simulation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulateResponse"
...
  /mcp:
    post:
      tags: [MCP]
      operationId: mcpJsonRpc
      summary: MCP JSON-RPC endpoint
      description: |
        JSON-RPC 2.0 methods:
        - `tools/list`
        - `tools/get`
        - `tools/call`

        For `tools/call`, timeout/idempotency can be provided either in `params` (`timeout_ms`, `idempotency_key`) or HTTP headers (`X-Timeout-Ms`, `Idempotency-Key`).
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/TimeoutMs"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JSONRPCRequest"
      responses:
        "200":
          description: JSON-RPC response envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSONRPCResponse"
...
components:
  securitySchemes:
    NexusApiKey:
      type: apiKey
      in: header
      name: X-NEXUS-GATEWAY-KEY
      description: Tenant API key (stored hashed in DB).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional JWT+JWKS auth mode when enabled by feature flags.
  parameters:
...
    XNexusActor:
      name: X-NEXUS-ACTOR
      in: header
      required: false
      schema:
        type: string
      description: Delegated actor identifier for audit/policy context.
    XNexusRole:
      name: X-NEXUS-ROLE
      in: header
      required: false
      schema:
        type: string
      description: Delegated role for policy context.
    XNexusScopes:
      name: X-NEXUS-SCOPES
      in: header
      required: false
      schema:
        type: string
      description: Comma-separated scopes; intersected with key/JWT scopes.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 255
      description: |
        Idempotency key for WRITE tool executions. `FAILED` records are terminal; retry requires a new key.
    TimeoutMs:
      name: X-Timeout-Ms
      in: header
      required: false
      schema:
        type: integer
        minimum: 1000
        maximum: 30000
      description: End-to-end timeout budget in milliseconds (clamped by server config).
...
  schemas:
    ErrorObject:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        error:
          $ref: "#/components/schemas/ErrorObject"
      required: [request_id, error]
    IdempotencyMeta:
      type: object
      properties:
        present:
          type: boolean
        outcome:
          type: string
          enum: [NEW, REPLAY, IN_PROGRESS, CONFLICT, MISSING_REQUIRED, SKIPPED_NOT_WRITE]
      required: [present, outcome]
    RunRequest:
      type: object
      properties:
        request_id:
          type: string
        tool_name:
          type: string
        input:
          type: object
          additionalProperties: true
        context:
          type: object
          additionalProperties: true
      required: [tool_name, input]
...
    RunSuccessResponse:
      type: object
      properties:
        request_id:
          type: string
        decision:
          type: string
          enum: [allow]
        tool_name:
          type: string
        status:
          type: string
          enum: [success]
        result:
          type: object
          additionalProperties: true
          nullable: true
        latency_ms:
          type: integer
        idempotency:
          $ref: "#/components/schemas/IdempotencyMeta"
      required: [request_id, decision, tool_name, status, latency_ms]
    RunBlockedResponse:
      type: object
      properties:
        request_id:
          type: string
        decision:
          type: string
          enum: [deny]
        status:
          type: string
          enum: [blocked]
        reason:
          type: string
        error:
          $ref: "#/components/schemas/ErrorObject"
        latency_ms:
          type: integer
        idempotency:
          $ref: "#/components/schemas/IdempotencyMeta"
      required: [request_id, decision, status, reason, error, latency_ms]
    RunErrorResponse:
      type: object
      properties:
        request_id:
          type: string
        decision:
          type: string
          enum: [allow]
        status:
          type: string
          enum: [error]
        error:
          $ref: "#/components/schemas/ErrorObject"
        latency_ms:
          type: integer
        idempotency:
          $ref: "#/components/schemas/IdempotencyMeta"
      required: [request_id, decision, status, error, latency_ms]
    SimulateRequest:
      allOf:
        - $ref: "#/components/schemas/RunRequest"
    SimulateResponse:
      type: object
      properties:
        request_id:
          type: string
        decision:
          type: string
          enum: [allow, deny]
        tool_name:
          type: string
        status:
          type: string
          enum: [success, error, blocked]
        reason:
          type: string
        error:
          $ref: "#/components/schemas/ErrorObject"
        explain:
          type: object
          additionalProperties: true
        latency_ms:
          type: integer
      required: [request_id, decision, tool_name, status, explain, latency_ms]
    Tool:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        kind:
          type: string
          enum: [http]
        description:
          type: string
          nullable: true
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        url:
          type: string
        input_schema:
          type: object
          additionalProperties: true
        output_schema:
          type: object
          additionalProperties: true
          nullable: true
        action_type:
          type: string
          enum: [read, write]
        classification:
          type: string
          enum: [internal, external]
        sensitivity:
          type: string
          enum: [low, medium, high]
        risk_level:
          type: integer
          minimum: 1
          maximum: 5
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - kind
        - method
        - url
        - input_schema
        - action_type
        - classification
        - sensitivity
        - risk_level
        - enabled
        - created_at
        - updated_at
    ToolCreateRequest:
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [http]
        description:
          type: string
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        url:
          type: string
        input_schema:
          type: object
          additionalProperties: true
        output_schema:
          type: object
          additionalProperties: true
        action_type:
          type: string
          enum: [read, write]
        classification:
          type: string
          enum: [internal, external]
          default: internal
        sensitivity:
          type: string
          enum: [low, medium, high]
          default: low
        risk_level:
          type: integer
          minimum: 1
          maximum: 5
        enabled:
          type: boolean
      required: [name, kind, method, url, input_schema, action_type, risk_level, enabled]
...
    ToolListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
      required: [items]
...
    JSONRPCRequest:
      type: object
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        id:
          type: string
        method:
          type: string
          enum: [tools/list, tools/get, tools/call]
        params:
          type: object
          additionalProperties: true
      required: [jsonrpc, id, method, params]
    JSONRPCError:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          additionalProperties: true
      required: [code, message]
    JSONRPCResponse:
      type: object
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        id:
          type: string
        result:
          type: object
          additionalProperties: true
        error:
          $ref: "#/components/schemas/JSONRPCError"
      required: [jsonrpc]
...